requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
pandas==2.2.3
python-dateutil==2.9.0
pyyaml==6.0.2
lxml==5.3.0
openpyxl==3.1.5
sqlalchemy==2.0.36
